Refactor safeTimestamp, delete ARM template discovery
